Add checking if document collection exists and return error if not

+++ WineLottery/Controllers/DraftController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.Documents;
 using Microsoft.Azure.Documents.Client;
 using Microsoft.Extensions.Configuration;
+using Newtonsoft.Json;
 
 namespace WineLottery.Controllers
 {
@@ -13,8 +14,8 @@ namespace WineLottery.Controllers
     public class DraftController : ControllerBase
     {
         private readonly IConfiguration config;
-        private DocumentClient dbClient;
-        protected internal string DbName => config["DbName"];
+        private readonly DocumentClient dbClient;
+        private string DbName => config["DbName"];
 
         public DraftController(IConfiguration config)
         {
@@ -35,6 +36,9 @@ namespace WineLottery.Controllers
         [HttpGet("{draftId}/Participants")]
         public ActionResult<IEnumerable<string>> Participants(string draftId)
         {
+            if (!CollectionExists(draftId))
+                return NotFound();
+
             IEnumerable<string> participants = dbClient.CreateDocumentQuery<Participant>(UriFactory.CreateDocumentCollectionUri(DbName, draftId))
                 .Where(d => !d.HasWon)
                 .Select(d => d.Name);
@@ -44,6 +48,9 @@ namespace WineLottery.Controllers
         [HttpGet("{draftId}/Winners")]
         public ActionResult<IEnumerable<string>> Winners(string draftId)
         {
+            if (!CollectionExists(draftId))
+                return NotFound();
+
             IEnumerable<string> participants = dbClient.CreateDocumentQuery<Participant>(UriFactory.CreateDocumentCollectionUri(DbName, draftId))
                 .Where(d => d.HasWon)
                 .Select(d => d.Name);
@@ -53,6 +60,9 @@ namespace WineLottery.Controllers
         [HttpPost("Participate")]
         public ActionResult Participant([FromBody]Participant participant)
         {
+            if (!CollectionExists(participant.DraftId))
+                return BadRequest();
+
             IEnumerable<Participant> participants = dbClient.CreateDocumentQuery<Participant>(UriFactory.CreateDocumentCollectionUri(DbName, participant.DraftId))
                 .Where(d => participant.UserId == d.UserId);
             if (participants.Any())
@@ -66,6 +76,9 @@ namespace WineLottery.Controllers
         [HttpGet("{draftId}/Winner")]
         public ActionResult<string> NextWinner(string draftId)
         {
+            if (!CollectionExists(draftId))
+                return NotFound();
+
             var participants = dbClient.CreateDocumentQuery<Participant>(UriFactory.CreateDocumentCollectionUri(DbName, draftId));
             int count = participants.Count();
             var random = new Random();
@@ -75,9 +88,18 @@ namespace WineLottery.Controllers
             return winner.Name;
         }
 
+        private bool CollectionExists(string draftId)
+        {
+            IQueryable<DocumentCollection> draft = dbClient.CreateDocumentCollectionQuery(UriFactory.CreateDatabaseUri(DbName)).Where(c => c.Id == draftId);
+            return draft.Any();
+        }
+
         [HttpDelete("{draftId}")]
         public ActionResult Delete(string draftId)
         {
+            if (!CollectionExists(draftId))
+                return NotFound();
+
             dbClient.DeleteDocumentCollectionAsync(UriFactory.CreateDocumentCollectionUri(DbName, draftId)).Wait();
             return Ok();
         }
@@ -85,10 +107,12 @@ namespace WineLottery.Controllers
 
     public class Participant
     {
+        [JsonIgnore]
         public string Id { get; set; }
         public string Name { get; set; }
         public string UserId { get; set; }
         public string DraftId { get; set; }
+        [JsonIgnore]
         public bool HasWon { get; set; }
     }
 }
